fix: align UT stubs and harden thread map access
- fix IAicpuTsThread::WriteValue stub signature to match uint16_t

- fix RtsqA5 UT call from WriteValue to WriteValue1B

- fix aclrtMalloc stub to return ACL_SUCCESS (aclError type)

- add wraparound-harmless comment for HcommRequestServiceOnThread s_msgId

- make ThreadEntity::DescribeAttr const

- guard g_ThreadMap accesses with mutex in alloc/service paths

+++ test/ut/stub/legacy/aicpu_ts_thread_interface.h
@@ -53,7 +53,7 @@ public:
         return HCCL_SUCCESS;
     }
 
-    HcclResult WriteValue(uint64_t addr, uint64_t value) const
+    HcclResult WriteValue(uint64_t addr, uint16_t value) const
     {
         return HCCL_SUCCESS;
     }
